requirements.txt
pillow
pyopengl
glfw
pyrr